+++ tests/testthat/test-run_plink.R
@@ -22,8 +22,6 @@ test_that("no duplicate --noweb flag", {
   )
 })
 
-
-
 test_that("warnings", {
   if (!is_plink_installed()) return()
   set.seed(314)
@@ -59,4 +57,47 @@ test_that("warnings", {
   file.remove(ped_filename)
   file.remove(map_filename)
   file.remove(phenotype_filename)
+})
+
+test_that("assoc_qt the PLINK way", {
+  skip("assoc_qt the PLINK way")
+  if (!is_plink_installed()) return()
+  set.seed(314)
+  assoc_qt_params <- create_demo_assoc_qt_params(
+    traits = rep(list(create_random_trait(n_snps = 1)), 2),
+    n_individuals = 10
+  )
+  ped_filename <- get_plinkr_tempfilename()
+  map_filename <- get_plinkr_tempfilename()
+  phenotype_filename <- get_plinkr_tempfilename()
+  save_ped_table_to_file(
+    ped_table = assoc_qt_params$ped_table,
+    ped_filename = ped_filename
+  )
+  save_map_table_to_file(
+    map_table = assoc_qt_params$map_table,
+    map_filename = map_filename
+  )
+  save_phenotype_table_to_file(
+    phenotype_table = assoc_qt_params$phenotype_table,
+    phenotype_filename = phenotype_filename
+  )
+  args <- c(
+    "--ped", ped_filename,
+    "--map", map_filename,
+    "--assoc",
+    "--pheno", phenotype_filename,
+    "--all-pheno"
+  )
+  expect_silent(run_plink(args))
+  qassoc_filenames <- paste0(
+    "plink.P",
+    seq_len(ncol(assoc_qt_params$phenotype_table) - 2),
+    ".qassoc"
+  )
+  expect_true(all(file.exists(qassoc_filenames)))
+  file.remove(ped_filename)
+  file.remove(map_filename)
+  file.remove(phenotype_filename)
+  file.remove(qassoc_filenames)
 })